separate Model component
